+++ app/javascript/components/Challenge/Challenge/Challenge.tsx
@@ -1,23 +1,19 @@
 import * as React from "react";
 import { useEffect, useState } from "react";
-import Grid from "@material-ui/core/Grid";
-import { Typography } from "@material-ui/core";
-import FormControl from "@material-ui/core/FormControl";
-import InputLabel from "@material-ui/core/InputLabel";
-import MenuItem from "@material-ui/core/MenuItem";
-import Select from "@material-ui/core/Select";
 import { RouteComponentProps, useParams } from "@reach/router";
 import Button from "@material-ui/core/Button";
 import { Partitions } from "../Partitions";
 import { ChallengeTable } from "../ChallengeTable";
 import { SubmissionDialog } from "../../SubmissionDialog/SubmissionDialog";
+import { VersionForm } from "../VersionForm";
+import Grid from "@material-ui/core/Grid";
+import { Typography } from "@material-ui/core";
 
 export const Challenge = (props: RouteComponentProps): JSX.Element => {
     const [open, setOpen] = useState(false);
 
     const [error, setError] = useState(null);
     const [version, setVersion] = useState<Version>();
-    const [name, setName] = useState<string>("1.0.0");
 
     const onOpen = () => {
         setOpen(true);
@@ -28,10 +24,6 @@ export const Challenge = (props: RouteComponentProps): JSX.Element => {
     };
 
     const parameters = useParams();
-
-    const onChange = (event: React.ChangeEvent<{ value: unknown }>) => {
-        setName(event.target.value as string);
-    };
 
     const init = {
         headers: {
@@ -65,19 +57,7 @@ export const Challenge = (props: RouteComponentProps): JSX.Element => {
                 </Grid>
 
                 <Grid item xs={3}>
-                    <FormControl fullWidth variant="outlined">
-                        <InputLabel id="version-label">Version</InputLabel>
-
-                        <Select
-                            id="version-select"
-                            label="Version"
-                            labelId="version-label"
-                            onChange={onChange}
-                            value={name}
-                        >
-                            <MenuItem value="1.0.0">1.0.0</MenuItem>
-                        </Select>
-                    </FormControl>
+                    <VersionForm />
                 </Grid>
             </Grid>
 
@@ -98,5 +78,3 @@ export const Challenge = (props: RouteComponentProps): JSX.Element => {
         </div>
     );
 };
-
-export default Challenge;
