+++ app/javascript/components/Challenge/VersionForm/VersionForm.css.ts
@@ -1,0 +1,7 @@
+import {createStyles, makeStyles} from "@material-ui/core/styles";
+
+const styles = {};
+
+const style = () => createStyles(styles);
+
+export const useStyles = makeStyles(style);
