+++ app/javascript/components/Challenge/VersionForm/index.ts
@@ -1,0 +1,1 @@
+export { VersionForm } from "./VersionForm";
